+++ exapp/nats.go
@@ -2,30 +2,27 @@ package exapp
 
 import (
 	"encoding/json"
-	"fmt"
 
 	model "github.com/edgexfoundry/edgex-go/pkg/models"
-	"github.com/mainflux/mainflux/logger"
-	"github.com/mteodor/edgex-app/events"
 	nats "github.com/nats-io/go-nats"
 )
 
 // NatsMSGHandler process incomming message
-func NatsMSGHandler(svc events.Service) nats.MsgHandler {
+func NatsMSGHandler(svc Service) nats.MsgHandler {
 
 	return func(msg *nats.Msg) {
-		logger.Info(fmt.Sprintf("Received a message: %s\n", string(msg.Data))
+		//logger.Info(fmt.Sprintf("Received a message: %s\n", string(msg.Data)))
 		processMsg(svc, msg)
 	}
 }
 
-func processMsg(svc events.Service, msg *nats.Msg) {
+func processMsg(svc Service, msg *nats.Msg) {
 	data := model.Event{}
 	json.Unmarshal(msg.Data, &data)
 
 	err := svc.RegisterEvent(data)
 	if err != nil {
-		logger.Info("failed to save data")
+		//logger.Error("failed to save data")
 	}
 
 }
