+++ exapp/postgres/events.go
@@ -12,8 +12,7 @@ import (
 
 	model "github.com/edgexfoundry/edgex-go/pkg/models"
 	"github.com/lib/pq"
-	"github.com/mainflux/mainflux/logger"
-	events "github.com/mteodor/edgex-app/events"
+	events "github.com/mteodor/edgex-app/exapp/events"
 )
 
 var _ events.EventsRepository = (*eventRepository)(nil)
@@ -32,7 +31,7 @@ func New(db *sql.DB) events.EventsRepository {
 
 func (rr eventRepository) Save(event model.Event) error {
 
-	logger.Info("saving event")
+	//logger.Info("saving event")
 
 	q := `INSERT INTO events (id, pushed, device, created, modifed, origin, event ) VALUES ($1, $2, $3, $4, $5, $6, $7)`
